Refactor dates into its own module
